atualização dos tipos int (32 bits) para 8 e 16 bits (de acordo com o contexto)

+++ imp/coletor_umidade_solo/recursos/SensorUmidadeSolo.cpp
@@ -1,6 +1,6 @@
 #include "SensorUmidadeSolo.h"
 
-SensorUmidadeSolo::SensorUmidadeSolo(int lacos, int intervalo)
+SensorUmidadeSolo::SensorUmidadeSolo(uint8_t lacos, uint16_t intervalo)
 {
 	this->lacos = lacos;
 	this->intervalo = intervalo;
